Working on login input/output
I'm trying to make sure that we are getting the values from the user. I'm having a problem with get the data from the API. I used some of the pseudo code but some times it would give me undefined on the page so I got rid of the if/else statements and alerts. I made a method that checks the inputs all at once for now, I would like to have it update everytime the user clicks away from the text box and a small message from that text box would appear to say whats wrong.