testing more inputs and outputs
I'm slowly getting the hang of it, I used some of the pseudocode and I assume I will probably look back at it some more. All I have right now is the thing telling me if I have the right information through console.log(). As for the actual submission of a new account, I haven't got to that yet.