Finished login page?
I got it so it looks like the way Thomas wanted, his code wasn't allowing me to goto different links. So I combined what I did with what he did and now it should take you to the link of the loginid. When registering a new user it is a bit slow for the database to update, I think this will be an issue in the future when we have multiple people signing on. For now though I am going to work on login keys so if the user logs off and so other people don't access some elses account the account is activated